pyproject.toml
[project]
name = "mediaref"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Suhwan Choi", email = "milkclouds00@gmail.com" }
]
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0",  # Requires v2 API (BaseModel, Field)
]

[project.optional-dependencies]
loader = [
    "numpy",
    "opencv-python",
    "pillow",
    "av",
    "requests",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"
